Crate tuxedo_parachain_runtime

The Tuxedo Template Runtime is an example runtime that uses most of the pieces provided in the wardrobe.

Modules

api

genesis

Helper module to build a genesis configuration for the template runtime.

opaque

Opaque types. These are used by the CLI to instantiate machinery that don’t need to know the specifics of the runtime. They can then be made to be agnostic over specific formats of data like extrinsics, allowing for them to continue syncing the network through upgrades to even the core data structures.

Structs

Runtime

The main struct in this module.

RuntimeApi

RuntimeApiImpl

Implements all runtime apis for the client side.

RuntimeParachainConfig

Enums

InnerConstraintChecker

A constraint checker is a piece of logic that can be used to check a transaction. For any given Tuxedo runtime there is a finite set of such constraint checkers. For example, this may check that input token values exceed output token values.

OuterVerifier

A verifier checks that an individual input can be consumed. For example that it is signed properly To begin playing, we will have two kinds. A simple signature check, and an anyone-can-consume check.

ParachainConstraintChecker

The Outer / Aggregate Constraint Checker for the Parachain runtime.

ParachainConstraintCheckerError

This type is generated by the #[tuxedo_constraint_checker] macro. It is a combined error type for the errors of each individual checker.

Constants

VERSION

This runtime version.

WASM_BINARY

WASM_BINARY_BLOATY

Functions

native_version

The version information used to identify this runtime when compiled natively.

Type Aliases

Block

Executive

Transaction